+++ mkchromecast/audio.py
@@ -40,7 +40,7 @@ In this block we check variables from __init__.py
 tray = mkchromecast.__init__.tray
 chunk_size = mkchromecast.__init__.chunk_size
 if debug == True:
-    print(':::ffmpeg::: chunk_size: ', chunk_size)
+    print(':::audio::: chunk_size: ', chunk_size)
 debug = mkchromecast.__init__.debug
 config = ConfigParser.RawConfigParser()
 configurations = config_manager()    # Class from mkchromecast.config
@@ -83,7 +83,7 @@ else:
         bitrate = ConfigSectionMap('settings')['bitrate']
         samplerate= ConfigSectionMap('settings')['samplerate']
         if debug == True:
-            print(':::ffmpeg::: tray ='+str(tray))
+            print(':::audio::: tray ='+str(tray))
             print(colors.warning('Configuration file exist'))
             print(colors.warning('Using defaults set there'))
             print(backend,codec,bitrate,samplerate)
@@ -117,8 +117,8 @@ else:
                 backends_dict[verifyif] = backend
                 backend = verifyif
                 if debug == True:
-                    print(':::ffmpeg::: Program '+str(backend)+' found in '+str(verifyif))
-                    print(':::ffmpeg::: backend dictionary '+str(backends_dict))
+                    print(':::audio::: Program '+str(backend)+' found in '+str(verifyif))
+                    print(':::audio::: backend dictionary '+str(backends_dict))
 
     if codec == 'mp3':
         appendmtype = 'mpeg'
@@ -491,7 +491,7 @@ else:
 app = Flask(__name__)
 
 if debug == True:
-    print(':::ffmpeg::: command '+str(command))
+    print(':::audio::: command '+str(command))
 
 @app.route('/')
 def index():