+++ mkchromecast/cast.py
@@ -292,8 +292,8 @@ class casting(object):
                 self.backend = ConfigSectionMap('settings')['backend']
 
         if self.backend == 'ffmpeg' or self.backend == 'avconv' or self.backend == 'parec':
-            import mkchromecast.ffmpeg
-            mtype = mkchromecast.ffmpeg.mtype
+            import mkchromecast.audio
+            mtype = mkchromecast.audio.mtype
             print(' ')
             print(colors.options('The media type string used is:')+' '+mtype)
             ncast.play_media('http://'+localip+':5000/stream', mtype)